pyproject.toml
[project]
name = "c-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bs4>=0.0.2",
    "gradio>=5.0.0",
    "langchain>=1.1.3",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.0.0",
    "langchain-groq>=1.1.1",
    "langchain-openai>=1.1.3",
    "langchain-tavily>=0.2.14",
    "langchain-text-splitters>=1.0.0",
    "langgraph>=1.0.5",
    "langgraph-cli[inmem]>=0.4.9",
    "python-dotenv>=1.2.1",
]
